+++ spec/spec_helper.rb
@@ -6,6 +6,13 @@ require 'webmock/rspec'
 
 WebMock.disable_net_connect!
 
+if !ENV['CODECOV_TOKEN'].nil? && !ENV['CODECOV_TOKEN'].empty?
+  require 'simplecov'
+  SimpleCov.start
+  require 'codecov'
+  SimpleCov.formatter = SimpleCov::Formatter::Codecov
+end
+
 RSpec.configure do |config|
   config.example_status_persistence_file_path = '.rspec_status'
   config.disable_monkey_patching!